pyproject.toml
[project]
name = "obsidian-vault-mcp"
version = "0.1.0"
description = "MCP server for semantic search over an Obsidian vault"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0",
    "lancedb>=0.6",
    "openai>=1.0",
    "ollama>=0.3",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "watchdog>=4.0",
    "pyyaml>=6.0",
    "pandas>=2.0",
]

[project.scripts]
vault-mcp = "src.server:main"
